fix: albums image pipeline
