+++ packages/musicbrainz/src/lib/ws-base-url.ts
@@ -1,0 +1,1 @@
+export const MB_WS_V2_BASE = 'https://musicbrainz.org/ws/2'
