+++ packages/musicbrainz/src/managers/manager.ts
@@ -1,0 +1,5 @@
+import type { MusicBrainzHttpClient } from '../http/musicbrainz-http-client.js'
+
+export abstract class Manager {
+  constructor(protected readonly http: MusicBrainzHttpClient) {}
+}
