+++ packages/musicbrainz/src/types/artist-credit.ts
@@ -1,0 +1,7 @@
+import type { IArtist } from './artist.js'
+
+export interface IArtistCredit {
+  artist: IArtist
+  joinphrase: string
+  name: string
+}
