+++ packages/musicbrainz/src/types/collection.ts
@@ -1,0 +1,9 @@
+import type { ITypedEntity } from './core.js'
+
+export interface ICollection extends ITypedEntity {
+  'type': string
+  'name': string
+  'recording-count': number
+  'editor': string
+  'entity-type': string
+}
