+++ packages/musicbrainz/src/types/index.ts
@@ -1,0 +1,29 @@
+export type {
+  IEntity,
+  ITypedEntity,
+  LifeSpan,
+  IPeriod,
+  IMatch,
+  ITag,
+  IUserVote,
+  IArtistTag,
+  IRating,
+  IGenre,
+  IArea,
+  IAlias,
+  IUrl,
+  RelationDirection,
+  IReleaseEvent,
+  ReleaseStatus,
+  ReleasePackaging,
+  ICoverArtArchive,
+} from './core.js'
+export type { IRelation, IMayHaveRelations } from './relations.js'
+export type { IArtistCredit } from './artist-credit.js'
+export type { IArtist } from './artist.js'
+export type { IRecording, IRecordingMatch, IRecordingList, ISearchResult } from './recording.js'
+export type { ITrack, IMedium } from './medium.js'
+export type { IReleaseGroup } from './release-group.js'
+export type { IRelease } from './release.js'
+export type { ICollection } from './collection.js'
+export type { CaaImage, CaaIndexJson, CaaThumbnails } from './caa.js'
