+++ packages/musicbrainz/src/types/release.ts
@@ -1,0 +1,31 @@
+import type {
+  ICoverArtArchive,
+  IEntity,
+  IReleaseEvent,
+  ReleasePackaging,
+  ReleaseStatus,
+} from './core.js'
+import type { IArtistCredit } from './artist-credit.js'
+import type { IMedium } from './medium.js'
+import type { IReleaseGroup } from './release-group.js'
+import type { IMayHaveRelations } from './relations.js'
+
+export interface IRelease extends IEntity, IMayHaveRelations {
+  'title': string
+  'text-representation'?: { language: string; script: string }
+  'disambiguation': string
+  'asin': null | string
+  'status': ReleaseStatus
+  'status-id': string
+  'packaging'?: ReleasePackaging
+  'release-events'?: IReleaseEvent[]
+  'date'?: string
+  'media'?: IMedium[]
+  'cover-art-archive'?: ICoverArtArchive
+  'country'?: string
+  'barcode'?: string
+  'artist-credit'?: IArtistCredit[]
+  'release-group'?: IReleaseGroup
+  'track-count'?: number
+  'count'?: number
+}
